ui/requirements.txt
streamlit
google-cloud-aiplatform==1.25.0